refactor code from index/flavor route to app/database files
